// types.ts
export interface FileSystemNode {
    type: 'file' | 'dir';
    content?: string; // content for text files
    appId?: string; // ID of the app to launch (for .exe)
    src?: string; // URL for media files or remote apps
    children?: { [key: string]: FileSystemNode }; // children for dirs
}